+++ webots_ros2_abb/launch/abb_irb4600.launch.py
@@ -27,13 +27,17 @@ from ament_index_python.packages import get_package_share_directory
 
 
 def generate_launch_description():
+    # Webots
     arguments = ['--mode=realtime', '--world=' +
                  os.path.join(get_package_share_directory('webots_ros2_abb'),
                               'worlds', 'abb_irb4600.wbt')]
     webots = launch_ros.actions.Node(package='webots_ros2_core', node_executable='webots_launcher',
                                      arguments=arguments, output='screen')
+    # Controller node
+    synchronization = launch.substitutions.LaunchConfiguration('synchronization', default=False)
     controller = ControllerLauncher(package='webots_ros2_abb',
                                     node_executable='abb_driver',
+                                    parameters=[{'synchronization': synchronization}],
                                     output='screen')
     return launch.LaunchDescription([
         webots,