+++ webots_ros2_universal_robot/launch/universal_robot_rviz.launch.py
@@ -29,24 +29,29 @@ from ament_index_python.packages import get_package_share_directory
 
 
 def generate_launch_description():
+    # Webots
     arguments = ['--mode=realtime', '--world=' +
                  os.path.join(get_package_share_directory('webots_ros2_universal_robot'),
                               'worlds', 'universal_robot_rviz.wbt')]
     webots = launch_ros.actions.Node(package='webots_ros2_core', node_executable='webots_launcher',
                                      arguments=arguments, output='screen')
+    # Controller nodes
+    synchronization = launch.substitutions.LaunchConfiguration('synchronization', default=False)
     URe5Controller = ControllerLauncher(package='webots_ros2_universal_robot',
                                         node_executable='universal_robot',
                                         # this argument should match the 'name' field
                                         # of the robot in Webots
                                         arguments=['--webots-robot-name=UR5e'],
+                                        parameters=[{'synchronization': synchronization}],
                                         output='screen')
     tfController = ControllerLauncher(package='webots_ros2_core',
                                       node_executable='tf_publisher',
                                       # this argument should match the 'name' field
                                       # of the robot in Webots
                                       arguments=['--webots-robot-name=tf_supervisor'],
+                                      parameters=[{'synchronization': synchronization}],
                                       output='screen')
-    # copy .rviz config file and update path ro URDF file.
+    # Copy .rviz config file and update path ro URDF file.
     templateRvizFile = os.path.join(get_package_share_directory('webots_ros2_ur_e_description'),
                                     'rviz', 'view_robot') + '.rviz'
     home = Path.home()
@@ -58,7 +63,7 @@ def generate_launch_description():
                                       get_package_share_directory('webots_ros2_ur_e_description'))
             with open(customRvizFile, 'w') as f2:
                 f2.write(content)
-    # rviz node
+    # Rviz node
     rviz = launch_ros.actions.Node(package='rviz2',
                                    node_executable='rviz2',
                                    arguments=['-d', customRvizFile],