+++ webots_ros2_core/CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package webots_ros2_core
 ------------------
 * Added support for multi robots.
 * Added a new tf_publisher node to publish tf of desired Webots nodes.
+* Added the possibility to run nodes in synchronized mode (using the 'synchronization' parameter).
 
 0.0.2 (2019-09-23)
 ------------------